+++ yenin/src/hw11/Lectures.java
@@ -1,10 +1,11 @@
-package onlineSchool;
+package hw11;
 
 import java.util.Arrays;
 import java.util.Scanner;
 
-public class Lectures extends SuperArray {
+public class Lectures {
     public static Lecture[] lectures = new Lecture[10];
+
     private int lectureCounter;
 
     public int getLectureCounter() {
@@ -15,28 +16,26 @@ public class Lectures extends SuperArray {
         return lectures;
     }
 
-    @Override
+
     public void getAll() {
         System.out.println(Arrays.toString(Lectures.getLectures()));
     }
 
-    @Override
     public void getById() {
         Scanner Scanner1 = new Scanner(System.in);
-        System.out.println(lectures[Scanner1.nextInt()]);
-        System.out.println();
+        int i = Scanner1.nextInt();
+        System.out.println(lectures[i]);
+        System.out.println(Persons.persons[lectures[i].getPersonId()]);
+
     }
 
-    @Override
     public void deleteById() {
         Scanner Scanner2 = new Scanner(System.in);
         lectures[Scanner2.nextInt()] = null;
     }
 
-    @Override
     public void add() {
         lectures[lectureCounter] = new Lecture(lectureCounter);
         lectureCounter++;
     }
 }
-
