+++ yenin/src/hw11/Main.java
@@ -1,0 +1,19 @@
+package hw11;
+
+import static hw11.Lectures.lectures;
+import static hw11.Persons.persons;
+public class Main {
+    public static void main(String[] args) {
+
+            Lectures lecture1 = new Lectures();
+            Persons persons1 = new Persons();
+            lecture1.add();
+            persons[0] = new Person(0);
+            persons[0].setRole(Role.TEACHER);
+            lectures[0].setPersonId(persons[0].getId());
+            lecture1.getById();
+
+
+    }
+}
+
