+++ yenin/src/hw11/Persons.java
@@ -1,0 +1,5 @@
+package hw11;
+
+public class Persons {
+    public static Person[] persons = new Person[10];
+}
