+++ yenin/src/hw11/Role.java
@@ -1,4 +1,4 @@
-package onlineSchool;
+package hw11;
 
 public enum Role {
     STUDENT,
