--- yenin/src/onlineSchool/Cource.java
@@ -1,7 +1,0 @@
-package onlineSchool;
-
-public class Cource extends Essence{
-    public Cource(int id) {
-        super(id);
-    }
-}
