--- yenin/src/onlineSchool/Essence.java
@@ -1,24 +1,0 @@
-package onlineSchool;
-
-public class Essence {
-    private int id;
-
-    public Essence(int id) {
-        this.id = id;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    @Override
-    public String toString() {
-        return "Essence{" +
-                "id=" + id +
-                '}';
-    }
-}
