--- yenin/src/onlineSchool/Student.java
@@ -1,7 +1,0 @@
-package onlineSchool;
-
-public class Student extends Essence {
-    public Student(int id) {
-        super(id);
-    }
-}
