--- yenin/src/onlineSchool/Teacher.java
@@ -1,7 +1,0 @@
-package onlineSchool;
-
-public class Teacher extends Essence{
-    public Teacher(int id) {
-        super(id);
-    }
-}
